refactor: creating a boiler to contains slices and changed the name of the store to index to avoid repetitions
