+++ src/hooks/useAuth.ts
@@ -5,7 +5,7 @@ import {
     startLoginWithEmailAndPassword,
     startLogout,
 } from '@/redux/thunks';
-import { AppDispatch } from '@/redux/store/store';
+import { AppDispatch } from '@/redux/store';
 import { useNavigate } from 'react-router-dom';
 import { useCheckAuth } from './useCheckAuth';
