+++ src/redux/slices/index.ts
@@ -1,0 +1,2 @@
+export { authSlice } from './authSlice';
+export { themeSlice } from './themeSlice';
