+++ src/types/index.d.ts
@@ -39,3 +39,16 @@ export interface AuthState {
     photoURL: string | null;
     errorMessage: string | null;
 }
+
+export type Theme = 'dark' | 'light' | 'system';
+
+export type ThemeProviderProps = {
+    children: React.ReactNode;
+    defaultTheme?: Theme;
+    storageKey?: string;
+};
+
+export type ThemeProviderState = {
+    theme: Theme;
+    setTheme: (theme: Theme) => void;
+};
